[server] Merge WorkspaceDeletionService into WorkspaceGC
